Rename module from openvpn_status to openvpn_status_log
